[LEIP-120] Refactor SettingsFragment and AppPreferences
