+++ ui/cyface/src/main/kotlin/de/cyface/app/CapturingFragment.kt
@@ -19,9 +19,7 @@
 package de.cyface.app
 
 import android.content.Intent
-import android.content.SharedPreferences
 import android.os.Bundle
-import android.preference.PreferenceManager
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -41,8 +39,8 @@ import de.cyface.app.dialog.ModalityDialog
 import de.cyface.app.ui.button.DataCapturingButton
 import de.cyface.app.utils.Map
 import de.cyface.app.utils.ServiceProvider
-import de.cyface.app.utils.SharedConstants.ACCEPTED_REPORTING_KEY
-import de.cyface.app.utils.SharedConstants.PREFERENCES_MODALITY_KEY
+import de.cyface.app.utils.capturing.settings.AppPreferences
+import de.cyface.camera_service.CameraPreferences
 import de.cyface.camera_service.CameraService
 import de.cyface.datacapturing.CyfaceDataCapturingService
 import de.cyface.datacapturing.persistence.CapturingPersistenceBehaviour
@@ -91,9 +89,14 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
         private set
 
     /**
-     * The `SharedPreferences` used to store the user's preferences.
+     * The `SharedPreferences` used to store the app preferences.
      */
-    private var preferences: SharedPreferences? = null
+    private lateinit var preferences: AppPreferences
+
+    /**
+     * The `SharedPreferences` used to store the camera preferences.
+     */
+    private lateinit var cameraPreferences: CameraPreferences
 
     /**
      * The `DataCapturingService` which represents the API of the Cyface Android SDK.
@@ -150,8 +153,7 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
             currentMeasurementsEvents = dataCapturingButton!!.loadCurrentMeasurementsEvents()
             map!!.render(currentMeasurementsTracks, currentMeasurementsEvents, false, ArrayList())
         } catch (e: NoSuchMeasurementException) {
-            val isReportingEnabled = preferences!!.getBoolean(ACCEPTED_REPORTING_KEY, false)
-            if (isReportingEnabled) {
+            if (preferences.getReportingAccepted()) {
                 Sentry.captureException(e)
             }
             Log.w(
@@ -184,7 +186,7 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
     ): View {
         _binding = FragmentCapturingBinding.inflate(inflater, container, false)
         dataCapturingButton = DataCapturingButton(this)
-        preferences = PreferenceManager.getDefaultSharedPreferences(requireContext())
+        preferences = AppPreferences(requireContext())
         // Register synchronization listener
         capturing.addConnectionStatusListener(this)
         syncButton = SynchronizationButton(capturing)
@@ -224,8 +226,7 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
 
     private fun showModalitySelectionDialogIfNeeded() {
         registerModalityTabSelectionListener()
-        val selectedModality = preferences!!.getString(PREFERENCES_MODALITY_KEY, null)
-        if (selectedModality != null) {
+        if (preferences.getModality() != null) {
             selectModalityTab()
             return
         }
@@ -242,7 +243,7 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
         val newModality = arrayOfNulls<Modality>(1)
         tabLayout.addOnTabSelectedListener(object : OnTabSelectedListener {
             override fun onTabSelected(tab: TabLayout.Tab) {
-                val oldModalityId = preferences!!.getString(PREFERENCES_MODALITY_KEY, null)
+                val oldModalityId = preferences.getModality()
                 val oldModality =
                     if (oldModalityId == null) null else Modality.valueOf(oldModalityId)
                 when (tab.position) {
@@ -253,9 +254,7 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
                     4 -> newModality[0] = Modality.TRAIN
                     else -> throw IllegalArgumentException("Unknown tab selected: " + tab.position)
                 }
-                preferences!!.edit()
-                    .putString(PREFERENCES_MODALITY_KEY, newModality[0]!!.databaseIdentifier)
-                    .apply()
+                preferences.saveModality(newModality[0]!!.databaseIdentifier)
                 if (oldModality != null && oldModality == newModality[0]) {
                     Log.d(
                         TAG,
@@ -305,7 +304,7 @@ class CapturingFragment : Fragment(), ConnectionStatusListener {
      */
     private fun selectModalityTab() {
         val tabLayout = binding.modalityTabs
-        val modality = preferences!!.getString(PREFERENCES_MODALITY_KEY, null)
+        val modality = preferences.getModality()
         Validate.notNull(modality, "Modality should already be set but isn't.")
 
         // Select the Modality tab
