+++ ui/cyface/src/main/kotlin/de/cyface/app/MainActivity.kt
@@ -26,14 +26,12 @@ import android.accounts.AuthenticatorException
 import android.accounts.OperationCanceledException
 import android.app.Activity
 import android.content.Intent
-import android.content.SharedPreferences
 import android.content.pm.PackageManager
 import android.location.Location
 import android.os.Bundle
 import android.os.Handler
 import android.os.Message
 import android.os.Parcel
-import android.preference.PreferenceManager
 import android.util.Log
 import android.widget.Toast
 import androidx.annotation.MainThread
@@ -54,11 +52,9 @@ import de.cyface.app.utils.Constants
 import de.cyface.app.utils.Constants.ACCOUNT_TYPE
 import de.cyface.app.utils.Constants.AUTHORITY
 import de.cyface.app.utils.ServiceProvider
-import de.cyface.app.utils.SharedConstants.ACCEPTED_REPORTING_KEY
-import de.cyface.app.utils.SharedConstants.DEFAULT_SENSOR_FREQUENCY
-import de.cyface.app.utils.SharedConstants.PREFERENCES_SENSOR_FREQUENCY_KEY
-import de.cyface.app.utils.SharedConstants.PREFERENCES_SYNCHRONIZATION_KEY
+import de.cyface.app.utils.capturing.settings.AppPreferences
 import de.cyface.camera_service.CameraListener
+import de.cyface.camera_service.CameraPreferences
 import de.cyface.camera_service.CameraService
 import de.cyface.camera_service.TriggerHandler
 import de.cyface.datacapturing.CyfaceDataCapturingService
@@ -120,9 +116,14 @@ class MainActivity : AppCompatActivity(), ServiceProvider, CameraServiceProvider
     private lateinit var navigation: NavController
 
     /**
-     * The `SharedPreferences` used to store the user's preferences.
+     * The `SharedPreferences` used to store the app preferences.
      */
-    private var preferences: SharedPreferences? = null
+    private lateinit var appPreferences: AppPreferences
+
+    /**
+     * The `SharedPreferences` used to store the camera preferences.
+     */
+    private lateinit var cameraPreferences: CameraPreferences
 
     /**
      * The authorization.
@@ -162,15 +163,13 @@ class MainActivity : AppCompatActivity(), ServiceProvider, CameraServiceProvider
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        preferences = PreferenceManager.getDefaultSharedPreferences(this)
+        appPreferences = AppPreferences(this)
+        cameraPreferences = CameraPreferences(this)
 
         // Location permissions are requested by MainFragment which needs to react to results
 
         // If camera service is requested, check needed permissions
-        val cameraEnabled = preferences!!.getBoolean(
-            de.cyface.camera_service.Constants.PREFERENCES_CAMERA_CAPTURING_ENABLED_KEY,
-            false
-        )
+        val cameraEnabled = cameraPreferences.getCameraEnabled()
         val permissionsMissing = ContextCompat.checkSelfPermission(
             this,
             Manifest.permission.CAMERA
@@ -194,8 +193,7 @@ class MainActivity : AppCompatActivity(), ServiceProvider, CameraServiceProvider
         }
 
         // Start DataCapturingService and CameraService
-        val sensorFrequency =
-            preferences!!.getInt(PREFERENCES_SENSOR_FREQUENCY_KEY, DEFAULT_SENSOR_FREQUENCY)
+        val sensorFrequency = appPreferences.getSensorFrequency()
         try {
             capturing = CyfaceDataCapturingService(
                 this.applicationContext,
@@ -331,8 +329,7 @@ class MainActivity : AppCompatActivity(), ServiceProvider, CameraServiceProvider
             capturing.shutdownDataCapturingService()
             // Before we only called: shutdownConnectionStatusReceiver();
         } catch (e: SynchronisationException) {
-            val isReportingEnabled = preferences!!.getBoolean(ACCEPTED_REPORTING_KEY, false)
-            if (isReportingEnabled) {
+            if (appPreferences.getReportingAccepted()) {
                 Sentry.captureException(e)
             }
             Log.w(TAG, "Failed to shut down CyfaceDataCapturingService. ", e)
@@ -382,8 +379,7 @@ class MainActivity : AppCompatActivity(), ServiceProvider, CameraServiceProvider
                     Validate.notNull(account)
 
                     // Set synchronizationEnabled to the current user preferences
-                    val syncEnabledPreference = preferences!!
-                        .getBoolean(PREFERENCES_SYNCHRONIZATION_KEY, true)
+                    val syncEnabledPreference = appPreferences.getUpload()
                     Log.d(
                         WiFiSurveyor.TAG,
                         "Setting syncEnabled for new account to preference: $syncEnabledPreference"
