+++ ui/cyface/src/main/kotlin/de/cyface/app/capturing/settings/SettingsViewModel.kt
@@ -21,6 +21,7 @@ package de.cyface.app.capturing.settings
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import de.cyface.app.utils.capturing.settings.AppPreferences
 
 /**
  * This is the [ViewModel] for the [SettingsFragment].
