+++ ui/cyface/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   android:orientation="vertical"
-  tools:context=".capturing.SettingsFragment">
+  tools:context=".capturing.settings.SettingsFragment">
 
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
