+++ ui/cyface/src/main/res/navigation/nav_graph.xml
@@ -29,7 +29,7 @@
 
     <fragment
       android:id="@+id/navigation_settings"
-      android:name="de.cyface.app.capturing.SettingsFragment"
+      android:name="de.cyface.app.capturing.settings.SettingsFragment"
       android:label="@string/settings"
       tools:layout="@layout/fragment_settings" />
   </navigation>
