+++ ui/digural/src/main/res/values-de/strings.xml
@@ -20,7 +20,7 @@
 
   <!-- Settings -->
   <string name="drawer_setting_speed_sensor">Externer Sensor</string>
-  <string name="camera_mode">Kameramodus</string>
+  <string name="camera_capturing">Kameraaufzeichnung</string>
   <string name="no_camera_available_toast">Es wurde keine Kamera gefunden die benutzt werden kann</string>
   <string name="bluetooth_setup_failed">Fehler beim Starten der Bluetooth Komponente</string>
   <string name="drawer_setting_camera_focus_distance">Kamera Fokus Distanz</string>
