+++ ui/digural/src/main/res/values-it/strings.xml
@@ -20,7 +20,7 @@
 
   <!-- Settings -->
   <string name="drawer_setting_speed_sensor">External sensor</string>
-  <string name="camera_mode">Camera mode</string>
+  <string name="camera_capturing">Camera capturing</string>
   <string name="no_camera_available_toast">Sorry, no camera found to be used.</string>
   <string name="bluetooth_setup_failed">Error starting the bluetooth component.</string>
   <string name="drawer_setting_camera_focus_distance">Camera focus distance</string>
