+++ ui/r4r/src/main/kotlin/de/cyface/app/r4r/capturing/settings/SettingsFragment.kt
@@ -26,6 +26,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import de.cyface.app.r4r.databinding.FragmentSettingsBinding
 import de.cyface.app.utils.ServiceProvider
+import de.cyface.app.utils.capturing.settings.AppPreferences
 import de.cyface.datacapturing.CyfaceDataCapturingService
 
 /**
