+++ ui/r4r/src/main/kotlin/de/cyface/app/r4r/capturing/settings/SettingsViewModelFactory.kt
@@ -20,6 +20,7 @@ package de.cyface.app.r4r.capturing.settings
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import de.cyface.app.utils.capturing.settings.AppPreferences
 
 /**
  * Factory which creates the [ViewModel] with the required dependencies.
