+++ utils/src/main/kotlin/de/cyface/app/utils/capturing/settings/CustomPreferences.kt
@@ -16,28 +16,30 @@
  * You should have received a copy of the GNU General Public License
  * along with the Cyface App for Android. If not, see <http://www.gnu.org/licenses/>.
  */
-package de.cyface.app.capturing.settings
+package de.cyface.app.utils.capturing.settings
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import de.cyface.app.utils.capturing.settings.AppPreferences
+import android.content.Context
+import androidx.core.content.edit
+import de.cyface.app.utils.trips.incentives.Incentives
+import de.cyface.utils.AppPreferences
 
 /**
- * Factory which creates the [ViewModel] with the required dependencies.
- *
- * Survives configuration changes and returns the right instance after Activity recreation.
+ * Preferences persisted with SharedPreferences.
  *
  * @author Armin Schnabel
  * @version 1.0.0
- * @since 3.4.0
+ * @since 7.8.1
  */
-class SettingsViewModelFactory(private val appPreferences: AppPreferences) :
-    ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(SettingsViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return SettingsViewModel(appPreferences) as T
+class CustomPreferences(context: Context) : AppPreferences(context) {
+
+    fun saveIncentivesUrl(incentivesUrl: String) {
+        preferences.edit {
+            putString(Incentives.INCENTIVES_ENDPOINT_URL_SETTINGS_KEY, incentivesUrl)
+            apply()
         }
-        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+
+    fun getIncentivesUrl(): String? {
+        return preferences.getString(Incentives.INCENTIVES_ENDPOINT_URL_SETTINGS_KEY, null)
     }
 }
